requirements.txt
httpx>=0.27.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
aiosqlite>=0.19.0
langdetect>=1.0.9
discord-webhook>=1.3.0
discord.py>=2.3.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
pydantic>=2.5.0
loguru>=0.7.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
